Search Users Issue Resolved

+++ src/pages/UserProfile.js
@@ -46,6 +46,8 @@ function UserProfile() {
 
   const [friends, setFriends] = useState([]);
 
+  const [editEvents, setEditEvents] = useState(false)
+
   let sortCategory = Array.isArray(isSelected) ? isSelected.sort() : [];
 
   // let sortCategory = [];
@@ -67,11 +69,6 @@ function UserProfile() {
       });
     }
   }, [loggedInUser?.id]);
-  
-  useEffect(()=>{
-    setLoggedInUser(loggedInUser)
-  },[loggedInUser?.id]);
-
 
   useEffect(() => {
     if (loggedInUser?.id) {
@@ -105,16 +102,25 @@ function UserProfile() {
     }
   }, [loggedInUser?.id]);
 
+  useEffect(() => {
+    setLoggedInUser(loggedInUser)
+  }, [loggedInUser?.id])
+
   function deleteMultiple() {
     const deleteEvent = userEvents
       .filter((events) => events.selected)
       .map((events) => events.event_id);
+      
+      
+    if (window.confirm("Are you sure you want to remove all selected events?")){
+      Promise.all(
+        deleteEvent.map((eventId) => {
+          axios.delete(`${API}/users/${loggedInUser?.id}/events/${eventId}`);
+        })
+        );
+      }
 
-    Promise.all(
-      deleteEvent.map((eventId) => {
-        axios.delete(`${API}/users/${loggedInUser?.id}/events/${eventId}`);
-      })
-    );
+    setEditEvents(false)
   }
 
   const acceptRequest = (senderId) => {
@@ -143,7 +149,7 @@ function UserProfile() {
       });
   };
 
-  // console.log(friends);
+console.log(isSelected)
 
   return (
     <>
@@ -153,19 +159,19 @@ function UserProfile() {
             <img
               src={loggedInUser?.profile_img}
               alt="profile-pic"
-              className="w-36 h-36"
+              className="w-36 h-36 basis-1/8 object-cover rounded"
             />
-            <div className="text-left w-1/6">
+            <div className="text-left basis-1/8">
               <h1>
                 <b>
                   {loggedInUser?.first_name} {loggedInUser?.last_name}{" "}
                   {loggedInUser?.pronouns ? (
                     <p className="inline">
                       <span>(</span>
-                      {loggedInUser?.pronouns}
+                        {loggedInUser?.pronouns} 
                       <span>)</span>
                     </p>
-                  ) : null}
+                    ) : null}
                 </b>
               </h1>
               <h2 className="text-emerald-500">@{loggedInUser?.username}</h2>
@@ -174,7 +180,7 @@ function UserProfile() {
                 {loggedInUser?.age?.age} years
               </h3>
             </div>
-            <div className="relative w-52">
+            <div className="relative w-52 basis-1/4 ml-5">
               <div className="align-middle inline">
                 <p className="text-left font-bold inline">Bio</p>
                 <BsPencilSquare
@@ -182,10 +188,10 @@ function UserProfile() {
                   className="inline text-cyan-800 cursor-pointer float-right mt-2"
                 />
               </div>
-              <section className="h-12 relative block">
-                <ImQuotesLeft className="text-orange-600 text-sm inline" />
-                <p className="px-4 inline">{loggedInUser?.bio}</p>
-                <ImQuotesRight className="text-orange-600 inline text-sm" />
+              <section className="w-52 h-12 relative flex flex-row">
+                <ImQuotesLeft className="text-orange-600 " />
+                <p className="px-4">{loggedInUser?.bio}</p>
+                <ImQuotesRight className="text-orange-600 " />
               </section>
             </div>
           </div>
@@ -200,6 +206,7 @@ function UserProfile() {
           ) : null}
         </div>
       </div>
+
       <div>
         {friendsRequest[0] &&
           friendsRequest.map((request) => {
@@ -261,38 +268,28 @@ function UserProfile() {
           />
         ) : null}
 
-        <fieldset
-          className={`w-3/4 border relative shadow-sm m-auto mb-8 ${
-            userEvents.length ? "h-52" : "h-20"
-          }`}
-        >
+        <fieldset className={`w-3/4 border relative shadow-sm m-auto mb-8 ${userEvents.length ? 'h-52' : 'h-20'}`}>
           <legend className="px-3 text-left ml-8">Events</legend>
           <div>
-            <div className="flex flex-wrap py-2 overflow-x-scroll h-44 4-full gap-y-8">
+            <div className="flex flex-wrap py-2 overflow-x-scroll h-44 gap-y-8">
               {Array.isArray(userEvents) && userEvents.length > 0 ? (
                 userEvents.map((event) => (
                   <div key={event.event_id}>
-                    <UserEvents event={event} />
+                    <UserEvents event={event} editEvents={editEvents} />
                   </div>
                 ))
-              ) : (
-                <p className="ml-7 py-2 text-gray-400">No events found.</p>
-              )}
+                ) : (
+                  <p className="ml-7 text-gray-400">No events found.</p>
+                  )}
+
             </div>
-            <button
-              onClick={() => navigate("/events")}
-              className="w-20 bg-blue-300 absolute right-3 top-3 rounded hover:bg-blue-200 shadow-md"
-            >
-              Add
-            </button>
-            {userEvents.length > 0 && (
               <button
-                onClick={deleteMultiple}
-                className="absolute right-3 bottom-3"
-              >
-                <BsTrash />
+                onClick={() => navigate("/events")}
+                className="w-20 bg-blue-300 absolute right-3 top-3 rounded hover:bg-blue-200 shadow-md"
+                >
+                Add
               </button>
-              {userEvents.length > 0 ? userEvents.length && !editEvents ? (
+              {userEvents.length > 0 && !editEvents ? (
                 <button
                   onClick={() => setEditEvents(!editEvents)}  
                   className="absolute right-3 bottom-3"
@@ -308,15 +305,11 @@ function UserProfile() {
                 >
                   <BsTrash />
                 </button>
-              ) : null }
+              )}
           </div>
         </fieldset>
 
-        <fieldset
-          className={`w-3/4 border relative shadow-sm m-auto ${
-            hostedEvents.length ? "h-40" : "h-20"
-          }`}
-        >
+        <fieldset className={`w-3/4 border relative shadow-sm m-auto ${hostedEvents.length ? 'h-52' : 'h-20' }`}>
           <legend className="px-3 text-left ml-8">Hosted Events</legend>
           <div className="flex flex-wrap px-3 py-2 overflow-y-auto">
             {hostedEvents.length > 0 ? (
@@ -325,10 +318,10 @@ function UserProfile() {
                   <UserHostedEvent hosted={hosted} />
                 </div>
               ))
-            ) : (
-              <p className="ml-5 py-2 text-gray-400">No hosted events found.</p>
-            )}
-          </div>
+              ) : (
+                <p className="ml-5 text-gray-400">No hosted events found.</p>
+                )}
+            </div>
           <button
             onClick={() => navigate("/events/new")}
             className="w-20 bg-blue-300 absolute right-3 top-3 rounded hover:bg-blue-200 shadow-md"
@@ -352,7 +345,9 @@ function UserProfile() {
                     alt="profile-pic"
                     className="w-10 h-15"
                   />
-                  <span className="ml-2">{friend.username}</span>
+                  <span className="ml-2">
+                    {friend.username}
+                  </span>
                 </Link>
               ))}
           </div>
